+++ keyboard_rewrite/test/test_helper.rb
@@ -1,9 +1,10 @@
+require "yast"
+require "yast/rspec"
+require_relative "helpers"
+
 SRC_PATH = File.expand_path("../../src", __FILE__)
 DATA_PATH = File.join(File.expand_path(File.dirname(__FILE__)), "data")
 ENV["Y2DIR"] = SRC_PATH
-
-require "yast"
-require "yast/rspec"
 
 if ENV["COVERAGE"]
   require "simplecov"
@@ -24,12 +25,7 @@ if ENV["COVERAGE"]
   end
 end
 
-def given_layouts(layouts_to_return)
-  allow(Cheetah).to receive(:run).with(
-    "localectl", "list-keymaps", stdout: :capture
-  ).and_return(layouts_to_return.join("\n"))
+RSpec.configure do |config|
+  config.include Helpers    # custom helpers
 end
 
-def mock_ui_events(*events)
-  allow(Yast::UI).to receive(:UserInput).and_return(*events)
-end
